move button dep within accordion
