+++ packages/gamut/src/AccordionButtonDeprecated/ButtonDeprecated/styles/mixins.scss
