+++ packages/gamut/src/AccordionButtonDeprecated/ButtonDeprecated/styles/variables.scss
