+++ packages/gamut/src/AccordionButtonDeprecated/ButtonDeprecatedBase/index.tsx
@@ -2,8 +2,8 @@ import cx from 'classnames';
 import { HTMLProps, ReactNode } from 'react';
 import * as React from 'react';
 
-import { ChildComponentDescriptor } from '../typings/react';
-import { omitProps } from '../utils/omitProps';
+import { ChildComponentDescriptor } from '../../typings/react';
+import { omitProps } from '../../utils/omitProps';
 // eslint-disable-next-line gamut/no-css-standalone
 import styles from './styles.module.scss';
